refactor : to dto
