+++ src/main/java/com/zuniorteam/racingcar/core/dto/StepResult.java
@@ -1,0 +1,22 @@
+package com.zuniorteam.racingcar.core.dto;
+
+import jdk.nashorn.internal.objects.annotations.Getter;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class StepResult {
+
+    private final List<Integer> carPositions;
+
+    public StepResult(List<Integer> carPositions) {
+        assert carPositions != null;
+
+        this.carPositions = new ArrayList<>(carPositions);
+    }
+
+    @Getter
+    public List<Integer> getCarPositions() {
+        return carPositions;
+    }
+}
