--- src/main/java/com/zuniorteam/racingcar/core/vo/StepResult.java
@@ -1,29 +1,0 @@
-package com.zuniorteam.racingcar.core.vo;
-
-import com.zuniorteam.racingcar.gui.drawer.PositionDrawer;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class StepResult {
-
-    private final List<Integer> carPositions;
-
-    public StepResult(List<Integer> carPositions) {
-        assert carPositions != null;
-
-        this.carPositions = new ArrayList<>(carPositions);
-    }
-
-    public String getDrawing(PositionDrawer positionDrawer) {
-        final StringBuilder stepResultString = new StringBuilder();
-
-        for (Integer position : carPositions) {
-            stepResultString
-                    .append(positionDrawer.draw(position))
-                    .append("\n");
-        }
-
-        return stepResultString.toString();
-    }
-}
